+++ backend/api_gateway/app/schemas/purchase_orders.py
@@ -18,6 +18,7 @@ Endpoints support:
 
 from pydantic import BaseModel, Field, field_validator, model_validator
 from typing import Optional, List, Dict, Any, Literal
+from decimal import Decimal
 from datetime import date
 
 
@@ -30,7 +31,7 @@ class PurchaseOrderItemCreate(BaseModel):
     item_id: Optional[str] = Field(None, description="Product UUID")
     item_code: Optional[str] = Field(None, max_length=50)
     description: str = Field(..., min_length=1, max_length=500)
-    quantity: float = Field(..., gt=0)
+    quantity: Decimal = Field(..., gt=0)
     unit: Optional[str] = Field(None, max_length=20)
     unit_price: int = Field(..., ge=0, description="Price per unit in IDR")
     discount_percent: float = Field(0, ge=0, le=100)
@@ -52,7 +53,7 @@ class PurchaseOrderItemUpdate(BaseModel):
     item_id: Optional[str] = None
     item_code: Optional[str] = None
     description: str = Field(..., min_length=1, max_length=500)
-    quantity: float = Field(..., gt=0)
+    quantity: Decimal = Field(..., gt=0)
     unit: Optional[str] = None
     unit_price: int = Field(..., ge=0)
     discount_percent: float = Field(0, ge=0, le=100)
